Re organize source code directory structure
