+++ src/components/ConfirmRemove/ConfirmRemove.tsx
